+++ services/src/main/java/org/keycloak/protocol/saml/EntityDescriptorDescriptionConverter.java
@@ -70,7 +70,7 @@ public class EntityDescriptorDescriptionConverter implements ClientDescriptionCo
 
     @Override
     public ClientRepresentation convertToInternal(String description) {
-        return loadEntityDescriptors(new ByteArrayInputStream(description.getBytes()));
+        return loadEntityDescriptors(new ByteArrayInputStream(description.getBytes()), new ClientRepresentation());
     }
 
     /**
@@ -148,7 +148,7 @@ public class EntityDescriptorDescriptionConverter implements ClientDescriptionCo
         return null;
     }
 
-    private static ClientRepresentation loadEntityDescriptors(InputStream is) {
+    public static ClientRepresentation loadEntityDescriptors(InputStream is, ClientRepresentation app) {
         Object metadata;
         try {
             metadata = SAMLParser.getInstance().parse(is);
@@ -169,26 +169,31 @@ public class EntityDescriptorDescriptionConverter implements ClientDescriptionCo
         }
 
         EntityDescriptorType entity = (EntityDescriptorType) entities.getEntityDescriptor().get(0);
-        String entityId = entity.getEntityID();
 
-        ClientRepresentation app = new ClientRepresentation();
-        app.setClientId(entityId);
+        if ( app.getClientId() == null)
+            app.setClientId(entity.getEntityID());
 
-        Map<String, String> attributes = new HashMap<>();
+        Map<String, String> attributes = app.getAttributes() == null ? new HashMap<>() : app.getAttributes();
         app.setAttributes(attributes);
 
         List<String> redirectUris = new LinkedList<>();
         app.setRedirectUris(redirectUris);
 
-        app.setFullScopeAllowed(true);
-        app.setProtocol(SamlProtocol.LOGIN_PROTOCOL);
-        attributes.put(SamlConfigAttributes.SAML_SERVER_SIGNATURE, SamlProtocol.ATTRIBUTE_TRUE_VALUE); // default to true
-        attributes.put(SamlConfigAttributes.SAML_SERVER_SIGNATURE_KEYINFO_EXT, SamlProtocol.ATTRIBUTE_FALSE_VALUE); // default to false
-        attributes.put(SamlConfigAttributes.SAML_SIGNATURE_ALGORITHM, SignatureAlgorithm.RSA_SHA256.toString());
-        attributes.put(SamlConfigAttributes.SAML_AUTHNSTATEMENT, SamlProtocol.ATTRIBUTE_TRUE_VALUE);
+        if ( app.getId() == null) {
+            //only during creation
+            app.setFullScopeAllowed(true);
+            app.setProtocol(SamlProtocol.LOGIN_PROTOCOL);
+            attributes.put(SamlConfigAttributes.SAML_SERVER_SIGNATURE, SamlProtocol.ATTRIBUTE_TRUE_VALUE); // default to true
+            attributes.put(SamlConfigAttributes.SAML_SERVER_SIGNATURE_KEYINFO_EXT, SamlProtocol.ATTRIBUTE_FALSE_VALUE); // default to false
+            attributes.put(SamlConfigAttributes.SAML_SIGNATURE_ALGORITHM, SignatureAlgorithm.RSA_SHA256.toString());
+            attributes.put(SamlConfigAttributes.SAML_AUTHNSTATEMENT, SamlProtocol.ATTRIBUTE_TRUE_VALUE);
+        }
+
         SPSSODescriptorType spDescriptorType = getSPDescriptor(entity);
         if (spDescriptorType.isWantAssertionsSigned()) {
             attributes.put(SamlConfigAttributes.SAML_ASSERTION_SIGNATURE, SamlProtocol.ATTRIBUTE_TRUE_VALUE);
+        } else {
+            attributes.put(SamlConfigAttributes.SAML_ASSERTION_SIGNATURE, SamlProtocol.ATTRIBUTE_FALSE_VALUE);
         }
         String logoutPost = getLogoutLocation(spDescriptorType, JBossSAMLURIConstants.SAML_HTTP_POST_BINDING.get());
         if (logoutPost != null) attributes.put(SamlProtocol.SAML_SINGLE_LOGOUT_SERVICE_URL_POST_ATTRIBUTE, logoutPost);
@@ -241,7 +246,8 @@ public class EntityDescriptorDescriptionConverter implements ClientDescriptionCo
                 attributes.put(ClientModel.POLICY_URI, spDescriptorType.getExtensions().getUIInfo().getPrivacyStatementURL().stream().filter(dn -> "en".equals(dn.getLang())).findFirst().orElse(spDescriptorType.getExtensions().getUIInfo().getPrivacyStatementURL().get(0)).getValue().toString());
             }
         }
-        
+
+        //TODO protocol mappers????
         app.setProtocolMappers(spDescriptorType.getAttributeConsumingService().stream().flatMap(att -> att.getRequestedAttribute().stream())
             .map(attr -> {
                 ProtocolMapperRepresentation mapper = new ProtocolMapperRepresentation();
@@ -258,6 +264,8 @@ public class EntityDescriptorDescriptionConverter implements ClientDescriptionCo
                 return mapper;
             }).collect(Collectors.toList()));
 
+        attributes.put(SamlConfigAttributes.SAML_CLIENT_SIGNATURE_ATTRIBUTE, SamlProtocol.ATTRIBUTE_FALSE_VALUE);
+        attributes.put(SamlConfigAttributes.SAML_ENCRYPT, SamlProtocol.ATTRIBUTE_FALSE_VALUE);
         for (KeyDescriptorType keyDescriptor : spDescriptorType.getKeyDescriptor()) {
             X509Certificate cert = null;
             try {