+++ services/src/main/java/org/keycloak/services/managers/RealmManager.java
@@ -46,6 +46,7 @@ import org.keycloak.protocol.oidc.OIDCLoginProtocol;
 import org.keycloak.protocol.oidc.OIDCLoginProtocolFactory;
 import org.keycloak.protocol.oidc.OIDCWellKnownProvider;
 import org.keycloak.protocol.oidc.mappers.AudienceResolveProtocolMapper;
+import org.keycloak.protocol.saml.SamlConfigAttributes;
 import org.keycloak.representations.idm.ApplicationRepresentation;
 import org.keycloak.representations.idm.ClientRepresentation;
 import org.keycloak.representations.idm.ClientScopeRepresentation;
@@ -56,6 +57,7 @@ import org.keycloak.representations.idm.RealmRepresentation;
 import org.keycloak.representations.idm.RoleRepresentation;
 import org.keycloak.services.scheduled.AutoUpdateIdentityProviders;
 import org.keycloak.services.scheduled.ClusterAwareScheduledTaskRunner;
+import org.keycloak.services.scheduled.ConfigureAutoUpdateSAMLClient;
 import org.keycloak.sessions.AuthenticationSessionProvider;
 import org.keycloak.services.clientregistration.policy.DefaultClientRegistrationPolicies;
 
@@ -581,6 +583,10 @@ public class RealmManager {
                 ClusterAwareScheduledTaskRunner taskRunner = new ClusterAwareScheduledTaskRunner(session.getKeycloakSessionFactory(), autoUpdateProvider, Long.parseLong(idp.getConfig().get(IdentityProviderModel.REFRESH_PERIOD)) * 1000);
                 timer.schedule(taskRunner, Long.parseLong(idp.getConfig().get(IdentityProviderModel.REFRESH_PERIOD)) * 1000, realm.getId()+"_AutoUpdateIdP_" + idp.getAlias());
             });
+        }
+        if (rep.getClients()!= null){
+            //saml autoupdated schedule task
+            rep.getClients().stream().filter(clientRep ->"saml".equals(clientRep.getProtocol()) && clientRep.getAttributes() != null && Boolean.valueOf(clientRep.getAttributes().get(SamlConfigAttributes.SAML_AUTO_UPDATED))).forEach(clientRep -> ConfigureAutoUpdateSAMLClient.configure(realm.getClientById(clientRep.getId()), realm, session));
         }
 
         setupClientServiceAccountsAndAuthorizationOnImport(rep, skipUserDependent);